minor mod to termstart

+++ bash/termStart.sh
@@ -1,13 +1,8 @@
 #! /bin/bash
 a=( `pgrep -f dropdown` )
-if [[ $? -eq 0 ]]; then
+err=$?
+terminator -H -b --layout=dropdown --profile=solar 
+if [[ ${err} -eq 0 ]]; then
 	zenity --title Killer --text "Kill existing drop downs?" --question >& /dev/null
-	if [[ $? -eq 0 ]]; then
-		terminator -b --layout=dropdown --profile=solar & 
-		kill -9 $a
-	else 
-		terminator -b --layout=dropdown --profile=solar &
-	fi
-else
-	terminator -b --layout=dropdown --profile=solar &
+	[[ $? -eq 0 ]] && kill -9 $a
 fi
